requirements.txt
asgiref==3.8.1
cffi==1.17.1
cryptography==43.0.3
Django==5.1.2
pycparser==2.22
PyJWT==2.9.0
sqlparse==0.5.1
